I added a new Database Project
after adding the datbase project called EMSDatabase.sqlproj I also added
a new directory called dbscripts and placed the sql script to create the
new schedule db provided by naveed in that folder. Now we can just
change that file whenever we want to modify the database. Also that
script will have to be executed on the DBs on our local machine when we
are working.